backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.0.2
python-dotenv==1.1.1
